In the case of slave nodes with intermittent network access (off/on multiple times), the BlockManagerMaster is some times found in a bad state (with zombie executor BlockManagerInfo), most likely due to some race condition. These zombie BlockManagerInfos will lead to exception (ClosedChannelException) when BlockManagerMaster is trying to remove Rdd/Shuffle/Broadcast, such as: java.io.IOException: Failed to send RPC 5807895775549752733 to /172.22.96.110:38256: java.nio.channels.ClosedChannelException The function checks the lastSeenMs of BlockManagerInfos and will expire them if they're zombie.
update comments: In the case of slave nodes with intermittent network access (off/on multiple times), the BlockManagerMasterEndpoint is some times found in a bad state (with zombie executor BlockManagerInfo), most likely due to some race condition. These zombie BlockManagerInfos will lead to exception (ClosedChannelException) when BlockManagerMasterEndpoint is trying to remove Rdd/Shuffle/Broadcast, such as: java.io.IOException: Failed to send RPC 5807895775549752733 to /172.22.96.110:38256: java.nio.channels.ClosedChannelException The function checks the lastSeenMs of BlockManagerInfos and will expire them if they're zombie.

revert the change in convertMapStatuses in MapOutputTracker, as status can't be null anymore (we set them to empty if task fails).